feat: Add @radix-ui/react-popover package
Add the "@radix-ui/react-popover" package to the frontend's package.json file. This package will be used to implement popovers in the UI components.

Also, create a new Popover component in the frontend/src/components/ui directory. This component will serve as the foundation for displaying popovers in the application.

The Popover component includes the following sub-components:
- PopoverTrigger: Responsible for triggering the display of the popover.
- PopoverContent: Renders the content of the popover.

These changes will enable the implementation of popovers in the application's UI.